chore: rendering the network by multiple components to distinguish between entity-data and the whole data
