+++ components/network.client.vue
@@ -1,13 +1,8 @@
 <script setup lang="ts">
-import Graph from "graphology";
-import circularpack from "graphology-layout/circlepack";
-import { DotIcon } from "lucide-vue-next";
+import type Graph from "graphology";
 import Sigma, { type Camera } from "sigma";
 import type { EdgeDisplayData, NodeDisplayData } from "sigma/types";
 import { nextTick, onMounted, ref } from "vue";
-
-import { networkConfig } from "@/config/network-visualisation.config";
-import { colors } from "@/project.config.json";
 
 interface State {
 	hoveredNode?: string;
@@ -21,11 +16,8 @@ interface State {
 	hoveredNeighbors?: Set<string>;
 }
 const props = defineProps<{
-	networkData: EntityFeature;
-	id: number;
+	graph: Graph;
 }>();
-
-const { getUnprefixedId } = useIdPrefix();
 
 interface NetworkContext {
 	graph: Graph;
@@ -34,67 +26,12 @@ interface NetworkContext {
 }
 
 const context: NetworkContext = {
-	graph: new Graph(),
+	graph: props.graph,
 	renderer: null,
 	camera: null,
 };
 
 const router = useRouter();
-const { entityColors } = colors;
-const defaultColor = "#666";
-let usedEntities: [string | undefined] = [""];
-
-watch(
-	() => {
-		return props.networkData;
-	},
-	(networkData) => {
-		/** Clear previous graph data. */
-		context.graph.clear();
-
-		/** Add source node. */
-		context.graph.addNode(props.id, {
-			label: networkData.properties.title,
-			color: getNodeColor(networkData.systemClass),
-			size: networkConfig.sourceNodeSize,
-		});
-
-		/** Add source node to agenda of nodes */
-
-		usedEntities.push(networkData.systemClass);
-
-		/** Add relations to target nodes. */
-		networkData.relations?.forEach((element) => {
-			if (element.relationTo == null) return;
-
-			const relationId = getUnprefixedId(element.relationTo);
-			const nodeClass = element.relationSystemClass;
-
-			if (nodeClass == null) return;
-
-			if (!usedEntities.includes(nodeClass)) {
-				usedEntities.push(nodeClass);
-			}
-			context.graph.addNode(relationId, {
-				label: element.label,
-				color: getNodeColor(nodeClass),
-				size: networkConfig.relationNodeSize,
-				url: element.relationTo,
-			});
-
-			context.graph.addEdge(props.id, relationId);
-		});
-
-		circularpack.assign(context.graph, { scale: 100 });
-	},
-	{ immediate: true },
-);
-
-function getNodeColor(nodeClass: string) {
-	// @ts-expect-error It's fine.
-	return entityColors[nodeClass] ?? defaultColor;
-}
-
 let hoverTimeOut: ReturnType<typeof setTimeout>;
 
 // let searchInput = "";
@@ -118,8 +55,6 @@ onMounted(async () => {
 	context.camera = context.renderer.getCamera();
 
 	context.renderer.on("clickNode", ({ node }) => {
-		console.log(node);
-
 		void router.push(node);
 	});
 
@@ -174,9 +109,7 @@ onMounted(async () => {
 });
 
 function setHoveredNode(node?: string) {
-	console.log(node);
 	if (node) {
-		console.log("entered.");
 		state.value.hoveredNode = node;
 		state.value.hoveredNeighbors = new Set(context.graph.neighbors(node));
 	} else {
@@ -226,7 +159,6 @@ function edgeReducer() {
 		const res: Partial<EdgeDisplayData> = { ...data };
 
 		if (state.value.hoveredNode && !context.graph.hasExtremity(edge, state.value.hoveredNode)) {
-			console.log(context.graph.hasExtremity(edge, state.value.hoveredNode));
 			res.hidden = true;
 		}
 
@@ -265,24 +197,5 @@ onScopeDispose(() => {
 </script>
 
 <template>
-	<div class="absolute z-10 m-3 flex w-full">
-		<Card class="w-max">
-			<span v-for="(color, entity) in entityColors" :key="entity">
-				<span v-if="usedEntities.includes(entity)" class="pr-4">
-					<DotIcon :size="50" :color="color" class="inline-block" />
-					<span>{{ entity }}</span>
-				</span>
-			</span>
-			<span v-for="entry in usedEntities" :key="entry">
-				<span
-					v-if="entry != null && entry !== '' && !Object.keys(entityColors).includes(entry)"
-					class="pr-4"
-				>
-					<DotIcon :size="50" :color="defaultColor" class="inline-block" />
-					<span>{{ entry }}</span>
-				</span>
-			</span>
-		</Card>
-	</div>
 	<div id="sigma-container" class="relative m-0 size-full overflow-hidden p-0"></div>
 </template>
